+++ bodhi/tests/server/scripts/test_check_policies.py
@@ -117,7 +117,9 @@ class TestCheckPolicies(BaseTestCase):
             self.assertEqual(result.exit_code, 0)
             update = self.db.query(models.Update).filter(models.Update.id == update.id).one()
             self.assertEqual(update.test_gating_status, models.TestGatingStatus.failed)
-            self.assertEqual(update.greenwave_summary_string, '1 of 2 tests are failed')
+            self.assertEqual(
+                update.greenwave_summary_string,
+                '1 of 2 tests are failed\n Missing: dist.depcheck')
 
         expected_query = {
             'product_version': 'fedora-17', 'decision_context': 'bodhi_update_push_stable',
@@ -220,3 +222,25 @@ class TestCheckPolicies(BaseTestCase):
                  'type': 'bodhi_update'}]}
         mock_greenwave.assert_called_once_with(config['greenwave_api_url'] + '/decision',
                                                expected_query)
+
+    @patch.dict(config, [('greenwave_api_url', 'http://domain.local')])
+    def test_archived_release_updates(self):
+        """Assert that updates for archived releases isn't being considered
+        by the script.
+        """
+        # Archive the F17 release
+        rel = self.db.query(models.Release).filter_by(name='F17').one()
+        rel.state = models.ReleaseState.archived
+        self.db.commit()
+
+        runner = testing.CliRunner()
+        update = self.db.query(models.Update).all()[0]
+        update.status = models.UpdateStatus.testing
+        self.db.commit()
+        with patch('bodhi.server.models.util.greenwave_api_post') as mock_greenwave:
+            mock_greenwave.side_effect = Exception(
+                'Greenwave should not be accessed for archived releases.')
+            result = runner.invoke(check_policies.check, [])
+            self.assertEqual(result.exit_code, 0)
+
+        self.assertEqual(mock_greenwave.call_count, 0)